Fixing issue with PageView nodes appearing in different sizes depending on contentScaleFactor.
The content size should be the same for the PageView node no matter what the contentScaleFactor is.
But when loading the texture from base64, the content size would depend on the scale factor.
This meant that the pageview nodes looked big in scale factors < 1, and normal only when scale factor = 1.