Adding a memory limit to the TextureCache
We would crash on Android because the TextureCache seemingly had no limit. It would keep acquiring memory to cache textures until the process was killed by the OS for using too much memory. Now we're able to set a limit, and if we're adding a new image but already above our target memory limit we'll drop any unused textures from the cache.